+++ overlays/karabiner-elements.nix
@@ -6,5 +6,7 @@ self: super: {
       inherit (old.src) url;
       hash = "sha256-gmJwoht/Tfm5qMecmq1N6PSAIfWOqsvuHU8VDJY8bLw=";
     };
+
+    dontFixup = true;
   });
 }